added api for comments to filmPresenter and added geting comments from serv, splited popup
